requirements.txt
# Azure OpenAI and OpenAI packages
openai==1.12.0
langchain-openai==0.0.8

# Core LangChain packages
langchain==0.1.14 
langchain-core==0.1.43
langchain-community==0.0.31 
langchain-chroma==0.1.0
langchainhub==0.1.21

# Supporting packages
pypdf==4.2.0  
beautifulsoup4==4.12.3 
httpx==0.23.3
langserve==0.0.30
chromadb==0.4.24 
faiss-cpu==1.8.0   
rapidocr-onnxruntime==1.3.16 
unstructured==0.13.2
fastapi==0.115.12
uvicorn==0.29.0
pydantic==1.10.13
sse_starlette==2.2.1
python-dotenv==1.0.0

# Transformers & model support
torch==2.2.2
transformers==4.50.3
accelerate==0.28.0
bitsandbytes==0.43.0 
huggingface-hub==0.30.2
sentence-transformers==2.6.1
